[Feat]1시간마다 랭킹 갱신, 랭킹 갱신에 따른 외형 추가, 유저 생성 시 기본 외형 추가, 유저 착용 외형 조회, 유저 착용 외형 변경 및 제약 조건 설정, 기타 API 변경 등

See merge request s10-final/S10P31C103!51

+++ BackEnd/src/main/java/com/ssafy/backend/domain/user/model/dto/request/UserUpdateRequest.java
@@ -17,6 +17,6 @@ public class UserUpdateRequest {
     private String account;
     private String password;
     private String nickname;
-    private long score;
-    private long ranking;
+    private int score;
+    private int ranking;
 }
